+++ Nvg.SMSService/SMSQuota/ISMSQuotaInteractor.cs
@@ -17,7 +17,7 @@ namespace Nvg.SMSService.SMSQuota
         /// <param name="channelKey">Channel Key</param>
         /// <param name="creditsUsed">Credits used for sending sms</param>
         /// <returns><see cref="SMSResponseDto{SMSQuotaDto}"/></returns>
-        SMSResponseDto<SMSQuotaDto> IncrementSMSQuota(string channelKey, int creditsUsed);
+        SMSResponseDto<SMSQuotaDto> IncrementSMSQuota(string channelKey, long creditsUsed);
 
         /// <summary>
         /// Check if SMS Quota of the channel is exceeded.Status is set to true or false based on value
