+++ Nvg.SMSService/SMSServiceProviders/SMSProviderResponse.cs
@@ -7,7 +7,7 @@ namespace Nvg.SMSService.SMSServiceProviders
     public class SmsProviderResponse
     {
         public string StatusMessage { get; set; }
-        public int? Unit { get; set; }
+        public long? Unit { get; set; }
         public long? SmsCost { get; set; }
     }
 }
